separate into people.js
