+++ javascript/three.js/src/people.js
@@ -1,0 +1,15 @@
+const people = [{
+  manager: 'Jen',
+  name: 'Bob'
+}, {
+  manager: 'Jen',
+  name: 'Sue'
+}, {
+  manager: 'Bob',
+  name: 'Shirley'
+}, {
+  manager: 'Bob',
+  name: 'Terrence'
+}]
+
+export default people
